section-2/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
requests
psycopg2-binary>=2.9
